requirements.txt
azure-ai-ml>=1.15.0
azure-identity>=1.15.0
openai>=1.30.0
